move get_target bed to RarediseaseAnalysisAPI
